// api/models/searchLibrary200.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Audiobookshelf API
 * Audiobookshelf API with autogenerated OpenAPI doc
 * OpenAPI spec version: 0.1.0
 */
import type { AuthorExpanded } from './authorExpanded';
import type { LibraryItemSearchResult } from './libraryItemSearchResult';
import type { SeriesBooks } from './seriesBooks';

export type SearchLibrary200 = {
  /** @nullable */
  book?: LibraryItemSearchResult[] | null;
  /**
   * TODO. Add podcast to libraryItemExpanded/media - try without oneOf
   * @nullable
   */
  podcast?: LibraryItemSearchResult[] | null;
  tags?: string[];
  authors?: AuthorExpanded[];
  series?: SeriesBooks[];
};